+++ tests/metadata.test.js
@@ -1,51 +1,19 @@
 const test = require("tape");
 const fs = require("fs");
-const jsyaml = require("js-yaml");
+const { readPost, buildList } = require("./utils");
 
-const readPost = (filename) => {
-  const buffer = fs.readFileSync(filename),
-    file = buffer.toString("utf8");
-
-  try {
-    const parts = file.split(/---\s*[\n^]/),
-      frontmatter = parts[1];
-
-    return {
-      name: filename,
-      file: file,
-      metadata: jsyaml.load(frontmatter),
-      content: parts[2],
-    };
-  } catch (err) {
-    console.log(
-      `\nCould not read metadata, check the syntax of the metadata and front matter in ${filename}`
-    );
-  }
-};
-
-const posts = fs.readdirSync(`_posts/`).reduce((arr, file) => {
-  if (file[0] !== ".") arr.push(`_posts/${file}`);
-  return arr;
-}, []);
-
-const testCases = fs.readdirSync(`test/`).reduce((arr, file) => {
-  if (file[0] !== ".") arr.push(`test/${file}`);
-  return arr;
-}, []);
-
-testCases.forEach((post) => {
+buildList("test").forEach((post) => {
   const file = readPost(post),
     metadata = file.metadata;
 
   test(post, (t) => {
     // has title
     t.ok(metadata.title, "has `title`");
-
     t.end();
   });
 });
 
-posts.forEach((post) => {
+buildList("_posts").forEach((post) => {
   const file = readPost(post),
     metadata = file.metadata;
